BookInventoryManagementSystem_python/requirements.txt
# Requirements
python-flask
python-mysql-connect
python-flask-sqlalchemy
python-pytest

